requirements.txt
Flask
requests
nmap
speedtest-cli
